server/requirements.txt
fastmcp[sse]>=2.7.0
langchain-ollama
uvicorn
fastapi
